# Conflicts:
#	products/ASC.Files/Core/Core/Dao/Interfaces/IFileDao.cs
#	products/ASC.Files/Server/Api/FoldersController.cs
#	products/ASC.Files/Server/Api/SettingsController.cs
#	products/ASC.Files/Server/Api/VirtualRoomsController.cs
#	web/ASC.Web.Api/Api/AuthenticationController.cs
#	web/ASC.Web.Api/Api/PortalController.cs
#	web/ASC.Web.Api/Api/SecurityController.cs
#	web/ASC.Web.Api/Api/Settings/MessageSettingsController.cs
#	web/ASC.Web.Api/Api/Settings/WebhooksController.cs
#	web/ASC.Web.Api/Api/Settings/WhitelabelController.cs

+++ products/ASC.Files/Service/Thumbnail/Builder.cs
@@ -24,59 +24,11 @@
 // content are licensed under the terms of the Creative Commons Attribution-ShareAlike 4.0
 // International. See the License terms at http://creativecommons.org/licenses/by-sa/4.0/legalcode
 
-using SixLabors.ImageSharp.Processing;
+
+using ASC.Data.Storage;
+using ASC.Data.Storage.DiscStorage;
 
 namespace ASC.Files.ThumbnailBuilder;
-
-[Singletone]
-public class BuilderQueue<T>
-{
-    private readonly ThumbnailSettings _config;
-    private readonly ILogger _logger;
-    private readonly IServiceScopeFactory _serviceScopeFactory;
-
-    public BuilderQueue(IServiceScopeFactory serviceScopeFactory, ILoggerProvider log, ThumbnailSettings settings)
-    {
-        _logger = log.CreateLogger("ASC.Files.ThumbnailBuilder");
-        _serviceScopeFactory = serviceScopeFactory;
-        _config = settings;
-    }
-
-    public async Task BuildThumbnails(IEnumerable<FileData<T>> filesWithoutThumbnails)
-    {
-        try
-        {
-            await using var scope = _serviceScopeFactory.CreateAsyncScope();
-
-            var fileDataProvider = scope.ServiceProvider.GetService<FileDataProvider>();
-
-            var premiumTenants = fileDataProvider.GetPremiumTenants();
-
-            filesWithoutThumbnails = filesWithoutThumbnails
-                .OrderByDescending(fileData => Array.IndexOf(premiumTenants, fileData.TenantId))
-                .ToList();
-
-
-            await Parallel.ForEachAsync(
-            filesWithoutThumbnails,
-            new ParallelOptions { MaxDegreeOfParallelism = _config.MaxDegreeOfParallelism },
-                async (fileData, token) =>
-            {
-                await using var scope = _serviceScopeFactory.CreateAsyncScope();
-                var commonLinkUtilitySettings = scope.ServiceProvider.GetService<CommonLinkUtilitySettings>();
-                commonLinkUtilitySettings.ServerUri = fileData.BaseUri;
-
-                var builder = scope.ServiceProvider.GetService<Builder<T>>();
-                await builder.BuildThumbnail(fileData);
-            }
-        );
-        }
-        catch (Exception exception)
-        {
-            _logger.ErrorBuildThumbnailsCount(filesWithoutThumbnails.Count(), exception);
-        }
-    }
-}
 
 [Scope]
 public class Builder<T>
@@ -91,10 +43,11 @@ public class Builder<T>
     private readonly GlobalStore _globalStore;
     private readonly PathProvider _pathProvider;
     private readonly IHttpClientFactory _clientFactory;
-    private readonly ThumbnailSettings _thumbnailSettings;
     private readonly SocketManager _socketManager;
     private readonly FFmpegService _fFmpegService;
     private readonly TempPath _tempPath;
+    private readonly TempStream _tempStream;
+    private IDataStore _dataStore;
 
     private readonly List<string> _imageFormatsCanBeCrop = new List<string>
             {
@@ -115,7 +68,7 @@ public class Builder<T>
         FFmpegService fFmpegService,
         TempPath tempPath,
         SocketManager socketManager,
-        ThumbnailSettings thumbnailSettings)
+        TempStream tempStream)
     {
         _config = settings;
         _tenantManager = tenantManager;
@@ -128,8 +81,8 @@ public class Builder<T>
         _clientFactory = clientFactory;
         _fFmpegService = fFmpegService;
         _tempPath = tempPath;
+        _tempStream = tempStream;
         _socketManager = socketManager;
-        _thumbnailSettings = thumbnailSettings;
         _globalStore = globalStore;
     }
 
@@ -138,6 +91,8 @@ public class Builder<T>
         try
         {
             _tenantManager.SetCurrentTenant(fileData.TenantId);
+
+            _dataStore = _globalStore.GetStore();
 
             var fileDao = _daoFactory.GetFileDao<T>();
             if (fileDao == null)
@@ -152,7 +107,7 @@ public class Builder<T>
         catch (Exception exception)
         {
             _logger.ErrorBuildThumbnailsTenantId(fileData.TenantId, exception);
-        }      
+        }
     }
 
     private async Task GenerateThumbnail(IFileDao<T> fileDao, FileData<T> fileData)
@@ -177,9 +132,7 @@ public class Builder<T>
                 return;
             }
 
-            var ext = FileUtility.GetFileExtension(file.Title);
-
-            if (!CanCreateThumbnail(ext) || file.Encrypted || file.RootFolderType == FolderType.TRASH || file.ContentLength > _config.AvailableFileSize)
+            if (!CanCreateThumbnail(file))
             {
                 file.ThumbnailStatus = Core.Thumbnail.NotRequired;
 
@@ -188,22 +141,24 @@ public class Builder<T>
                 return;
             }
 
+            await fileDao.SetThumbnailStatusAsync(file, Core.Thumbnail.Creating);
+
+            var ext = FileUtility.GetFileExtension(file.Title);
+
             if (IsVideo(ext))
             {
                 await MakeThumbnailFromVideo(fileDao, file);
             }
+            else if (IsImage(ext))
+            {
+                await MakeThumbnailFromImage(fileDao, file);
+            }
             else
             {
-
-                if (IsImage(ext))
-                {
-                    await CropImage(fileDao, file);
-                }
-                else
-                {
-                    await MakeThumbnail(fileDao, file);
-                }
+                await MakeThumbnailFromDocs(fileDao, file);
             }
+
+            await fileDao.SetThumbnailStatusAsync(file, Core.Thumbnail.Created);
 
             var newFile = await fileDao.GetFileStableAsync(file.Id);
 
@@ -228,26 +183,36 @@ public class Builder<T>
         var thumbPath = _tempPath.GetTempFileName("jpg");
         var tempFilePath = _tempPath.GetTempFileName(Path.GetExtension(file.Title));
 
-        using (var fileStream = new FileStream(tempFilePath, FileMode.Open, FileAccess.ReadWrite, System.IO.FileShare.Read))
+        try
         {
-            await streamFile.CopyToAsync(fileStream);
+            using (var fileStream = new FileStream(tempFilePath, FileMode.Open, FileAccess.ReadWrite, System.IO.FileShare.Read))
+            {
+                await streamFile.CopyToAsync(fileStream);
+            }
+
+            await _fFmpegService.CreateThumbnail(tempFilePath, thumbPath);
+
+            using (var streamThumb = new FileStream(thumbPath, FileMode.Open, FileAccess.ReadWrite, System.IO.FileShare.Read))
+            {
+                await Crop(fileDao, file, streamThumb);
+            }
         }
-
-        await _fFmpegService.CreateThumbnail(tempFilePath, thumbPath);
-
-        using (var streamThumb = new FileStream(thumbPath, FileMode.Open, FileAccess.ReadWrite, System.IO.FileShare.Read))
+        finally
         {
-            await Crop(fileDao, file, streamThumb);
-        }
+            if (Path.Exists(thumbPath))
+            {
+                File.Delete(thumbPath);
+            }
 
-        File.Delete(thumbPath);
-        File.Delete(tempFilePath);
+            if (Path.Exists(tempFilePath))
+            {
+                File.Delete(tempFilePath);
+            }
+        }
     }
 
-    private async Task MakeThumbnail(IFileDao<T> fileDao, File<T> file)
+    private async Task MakeThumbnailFromDocs(IFileDao<T> fileDao, File<T> file)
     {
-        await fileDao.SetThumbnailStatusAsync(file, Core.Thumbnail.Creating);
-       
         foreach (var w in _config.Sizes)
         {
             _logger.DebugMakeThumbnail1(file.Id.ToString());
@@ -261,7 +226,7 @@ public class Builder<T>
                 try
                 {
                     (resultPercent, thumbnailUrl) = await GetThumbnailUrl(file, _global.DocThumbnailExtension.ToString(), w.Width, w.Height);
-                    
+
                     if (resultPercent == 100)
                     {
                         break;
@@ -287,7 +252,7 @@ public class Builder<T>
                         {
                             _logger.WarningMakeThumbnail(file.Id.ToString(), thumbnailUrl, resultPercent, attempt, exception);
                         }
-                    } 
+                    }
                     else
                     {
                         _logger.WarningMakeThumbnail(file.Id.ToString(), thumbnailUrl, resultPercent, attempt, exception);
@@ -308,11 +273,8 @@ public class Builder<T>
             }
             while (string.IsNullOrEmpty(thumbnailUrl));
 
-            await SaveThumbnail(fileDao, file, thumbnailUrl, w.Width, w.Height);
+            await SaveThumbnail(fileDao, file, thumbnailUrl, w.Width, w.Height, w.ResizeMode, AnchorPositionMode.Top);
         }
-
-        await fileDao.SetThumbnailStatusAsync(file, Core.Thumbnail.Created);
-
     }
 
     private async Task<(int, string)> GetThumbnailUrl(File<T> file, string toExtension, int width, int height)
@@ -355,7 +317,7 @@ public class Builder<T>
         return (operationResultProgress, url);
     }
 
-    private async Task SaveThumbnail(IFileDao<T> fileDao, File<T> file, string thumbnailUrl, int width, int height)
+    private async Task SaveThumbnail(IFileDao<T> fileDao, File<T> file, string thumbnailUrl, int width, int height, ResizeMode resizeMode, AnchorPositionMode anchorPositionMode = AnchorPositionMode.Center)
     {
         _logger.DebugMakeThumbnail3(file.Id.ToString(), thumbnailUrl);
 
@@ -368,11 +330,22 @@ public class Builder<T>
         {
             using (var sourceImg = await Image.LoadAsync(stream))
             {
-                await CropAsync(sourceImg, fileDao, file, width, height);
+                await CropAsync(sourceImg, fileDao, file, width, height, resizeMode, anchorPositionMode);
             }
         }
 
         _logger.DebugMakeThumbnail4(file.Id.ToString());
+    }
+
+    public bool CanCreateThumbnail(File<T> file)
+    {
+        var ext = FileUtility.GetFileExtension(file.Title);
+
+        if (!CanCreateThumbnail(ext) || file.Encrypted || file.RootFolderType == FolderType.TRASH) return false;
+        if (IsVideo(ext) && file.ContentLength > _config.MaxVideoFileSize) return false;
+        if (IsImage(ext) && file.ContentLength > _config.MaxImageFileSize) return false;
+
+        return true;
     }
 
     private bool CanCreateThumbnail(string extention)
@@ -390,7 +363,7 @@ public class Builder<T>
         return _fFmpegService.ExistFormat(extention);
     }
 
-    private async Task CropImage(IFileDao<T> fileDao, File<T> file)
+    private async Task MakeThumbnailFromImage(IFileDao<T> fileDao, File<T> file)
     {
         _logger.DebugCropImage(file.Id.ToString());
 
@@ -404,36 +377,37 @@ public class Builder<T>
 
     private async Task Crop(IFileDao<T> fileDao, File<T> file, Stream stream)
     {
-        using (var sourceImg = await Image.LoadAsync(stream))
+        using var sourceImg = await Image.LoadAsync(stream);
+
+        if (_dataStore is DiscDataStore)
         {
-            //var tasks = new List<Task>();
-
-            //foreach (var w in config.Sizes)
-            //{
-            //    tasks.Add(CropAsync(sourceImg, fileDao, file, w.Width, w.Height));
-            //}
-
-            //await Task.WhenAll(tasks.ToArray());
-
-            //await Parallel.ForEachAsync(config.Sizes, (w, b) => CropAsync(sourceImg, fileDao, file, w.Width, w.Height));
-
-            await fileDao.SetThumbnailStatusAsync(file, Core.Thumbnail.Creating);
-
             foreach (var w in _config.Sizes)
             {
-                await CropAsync(sourceImg, fileDao, file, w.Width, w.Height);
+                await CropAsync(sourceImg, fileDao, file, w.Width, w.Height, w.ResizeMode);
             }
-
-            await fileDao.SetThumbnailStatusAsync(file, Core.Thumbnail.Created);
+        }
+        else
+        {
+            await Parallel.ForEachAsync(_config.Sizes, new ParallelOptions { MaxDegreeOfParallelism = 3 }, async (w, t) =>
+            {
+                await CropAsync(sourceImg, fileDao, file, w.Width, w.Height, w.ResizeMode);
+            });
         }
 
-        GC.Collect();
+        //        GC.Collect();
     }
 
-    private async ValueTask CropAsync(Image sourceImg, IFileDao<T> fileDao, File<T> file, int width, int height)
+    private async ValueTask CropAsync(Image sourceImg,
+                                      IFileDao<T> fileDao,
+                                      File<T> file,
+                                      int width,
+                                      int height,
+                                      ResizeMode resizeMode,
+                                      AnchorPositionMode anchorPositionMode = AnchorPositionMode.Center)
     {
-        using var targetImg = GetImageThumbnail(sourceImg, width);
-        using var targetStream = new MemoryStream();
+        using var targetImg = GetImageThumbnail(sourceImg, width, height, resizeMode, anchorPositionMode);
+        using var targetStream = _tempStream.Create();
+
         switch (_global.ThumbnailExtension)
         {
             case ThumbnailExtension.bmp:
@@ -462,11 +436,20 @@ public class Builder<T>
                 break;
         }
 
-        await _globalStore.GetStore().SaveAsync(fileDao.GetUniqThumbnailPath(file, width, height), targetStream);     
+        await _dataStore.SaveAsync(fileDao.GetUniqThumbnailPath(file, width, height), targetStream);
     }
 
-    private Image GetImageThumbnail(Image sourceBitmap, int thumbnaillWidth)
+    private Image GetImageThumbnail(Image sourceBitmap, int thumbnaillWidth, int thumbnaillHeight, ResizeMode resizeMode, AnchorPositionMode anchorPositionMode)
     {
-        return sourceBitmap.Clone(x => x.BackgroundColor(Color.White).Resize(thumbnaillWidth, 0));
+        return sourceBitmap.Clone(x =>
+        {
+            x.Resize(new ResizeOptions
+            {
+                Size = new Size(thumbnaillWidth, thumbnaillHeight),
+                Mode = resizeMode,
+                Position = anchorPositionMode,
+                Sampler = KnownResamplers.Lanczos8
+            });
+        });
     }
 }
